disable some modules in the maude backend for the time being (#3213)
These modules contain lemmas that are needed for the Haskell backend,
but, due to the fact that the Maude backend implements full matching
modulo AC on integer addition and map concatenation, cause the Maude
backend to not terminate. We will eventually create separate versions
of these modules for the Maude backend as needed in the future.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>