New KCFGShow class for upstreamed KCFG display (https://github.com/runtimeverification/pyk/pull/258)
This makes a new module `kcfg/show.py` with class `KCFGShow` which is
responsible for displaying KCFG results.

- Module `kcfg/show.py` is added, with class `KCFGShow`.
- Functionality `KCFGShow.show` is imported from KEVM foundry
integration, which enables displaying the CFG as text, but with
additional information that is generally useful, including:
  - A list of nodes to display in full,
  - A list of node-pairs to display state deltas for,
  - Whether to minimize the generated output or not,
- An option to supply a node_printer function to incorporate into the
KCFG display output.
- Functionality `KCFGShow.dump` is imported from KEVM foundry
integration, which enables dumping hash-indexed files with pretty
variants of each node/edge/cover in the graph. Extra option in:
  - `dot`: if specified, will also dump a DOT rendering of the graph.

---------

Co-authored-by: devops <devops@runtimeverification.com>