Turn implication check validity back on (https://github.com/runtimeverification/pyk/pull/261)
Fixes: https://github.com/runtimeverification/pyk/issues/248

KEVM had an issue where it was sending macro symbols to the SMT checker,
which had undesirable effects. That has been fixed here:
https://github.com/runtimeverification/evm-semantics/pull/1627

This PR:

- Re-enables the check that if we get an unsatisfiable implication, then
the substitution/predicate are empty.
- Adds another check that if we get a satisfiable implication, that the
substitution/predicate are not empty, emits debug message if not the
case. This can be useful for cases where the backend is able to tell
that implication is not holding, but you want to get more information
about why as the user (because it returns the unificand if it doesn't).
- Downgrades an info message to a debug message that was not very
useful.
- Disables the (now useless/redundant) check that the instantiated
consequent is `#Bottom`, which should make implication checker somewhat
faster.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>