Introduce Constrained Substitution, CSubst, class (#294)
This is somewhat a blocker on:
https://github.com/runtimeverification/pyk/pull/124

Working with that PR, it became clear that the `Cover` and `Split` nodes
have some duplicate data. So here we factor out that data/functionality
into the `CSubst` class, which stores a substitution and a constraint.

- Class `CSubst` is added, with functionality:
  - `apply`: apply a `CSubst` to a `CTerm` to get a new `CTerm`.
- `add_constraint`: return a new `CSubst` with the additional
constraint.
- Method `CTerm.match_with_constraint` now returns an `Optional[CSubst]`
instead.
- Adds methods `CTerm.{to_dict,from_dict}` and
`CSubst.{to_dict,from_dict}`, and uses those directly in the various
`KCFG.{to_dict,from_dict}` places.
- Removes a method `path_condition` from `kcfg.py` which had no uses
here.
- Stores a `CSubst` directly as the data of a `KCFG.Cover` instead of
storing a constraint/substitution directly.
- Adjust `KCFGExplore.cterm_implies` to return a `CSubst` instead of a
tuple of `Subst, KInner`.
- Adjusts the tests to the new datastructures.
- Increases the parallelism of tests on CI.

---------

Co-authored-by: devops <devops@runtimeverification.com>