Direct storage of split nodes (https://github.com/runtimeverification/pyk/pull/124)
~Blocked on: https://github.com/runtimeverification/pyk/pull/122~
~Blocked on: https://github.com/runtimeverification/pyk/pull/294~
~Blocked on: https://github.com/runtimeverification/pyk/pull/302~
~Blocked on: https://github.com/runtimeverification/pyk/pull/303~

This stores Split nodes as their own data-structure, rather than
shimming them in as edges/covers. As I've been working on the summarizer
paper, it's apparent that we need more restrictions on what a valid KCFG
is. Explicit representation of split edges allows us to do that.

**NOTE**: We _lose_ the ability to represent non-determinism properly
here! This is because we are restricting each node to having _only one_
outgoing Edge, Cover, or Split from it! We can recover this later with
direct representation of non-determinism, as a multi-edge that takes
exactly one step on each edge to the non-deterministic transitions.

- Add `KCFG.Successor` class, which can be an edge or a multi-edge, and
just has the `source` attribute. Make `KCFG.EdgeLike` a subclass of this
class.
- Add `KCFG.Split` class, which is a subclass of `KCFG.Successor`, and
stores a list of `targets`. Each `target` is a `Tuple[Node, CSubst]`,
which is basically a "reverse cover", it points at the more specialized
node instead of the more abstract node.
- Stricter enforcement on the structure of the `KCFG`. In particular,
every node can have at _most_ one successor. We were allowing multiple
0-step `KCFG.Edge` out of a given node to represent case splits. Those
are directly represented as the `KCFG.Split` multi-edge now.
- Adjusts the various routines to use the new `KCFG.split`
datastructure, including:
  - `reachable_nodes`: traverses splits forward/backwards.
- `paths_between`: almost complete rewrite now that we must support
multi-edges.
- `KCFG.pretty_segments` now directly detects splits instead of
implicitly detecting them as "multiple outgoing 0-depth `KCFG.Edge` on
the same node".
- Updates to `KCFGExplore` routines:
- `KCFGExplore.step` is now allowed to assume only one successor edge,
simplifying its logic.
- `KCFGExplore.all_path_reachability_prove` can now take advantage of
new `KCFG.split_on_constraints`, which stores the split as a
`KCFG.Split` node.
- Adds a test of reachable nodes in the reverse case, that was failing
before fixing a bug in the handling of `KCFG.Split` nodes.

---------

Co-authored-by: devops <devops@runtimeverification.com>