Correctly handle cut-rule case in AGProver (https://github.com/runtimeverification/pyk/pull/341)
Currently, when the user supplies a `cut-rule` to the `AGProver`, it
will have a hard time making progress. This is because a cut-rule
returns the state _prior_ to the rule, and then the prover tries to
resume progress from that state, and halts again because it sees the
same cut-rule. Then the `AGProver` thinks it's stuck because it cannot
make any progress on that term.

Cut-rules are needed to identify loop-heads for bounded model checking
(part of https://github.com/runtimeverification/pyk/issues/310), and for
the summarizer.

This PR:

- Introduces 2 tests that fail currently with the `AGProver`. One of
them applies a cut-rule immediately on the first step, but should
succeed on the state following the first step. The other applies the cut
rule a few steps in, but should succeed immediately following the cut
rule application.
- Changes the `KCFGExplore.cterm_execute` to pass through the results
that come back from 0-step cut-rules, rather than removing that case.
- Changes the `AGProver` to detect if we stopped because of a 0-step
cut-rule application, and if so continues execution from the state
_following_ the cut-rule application.
- Changes the tests in `tests/integration/kcfg/` of the `*Prover`
classes to assert that `proof.status == ProofStatus.PASSED` instead of
inspecting the frontier/stuck states directly.
- Fixes a minor bug in printing out KCFGs where we were considering a
node as `prior_on_trace` too early.

---------

Co-authored-by: devops <devops@runtimeverification.com>