Added integration test for lnternalTerm.__str__ binding (#346)
As noted in runtimeverification/llvm-backend#724, there was previously a
bug in the InternalTerm.\_\_str\_\_ binding which led to extraneous data
being included in the string, causing a crash.

This issue wasn't reproducible in the LLVM backend testing environment,
so I am adding a test here on the Pyk side instead. I confirmed that the
test fails if the LLVM backend fix is rolled back.

---------

Co-authored-by: devops <devops@runtimeverification.com>