AGBMCProver: implement basic bounded-model-checker prover (https://github.com/runtimeverification/pyk/pull/347)
Part of: https://github.com/runtimeverification/pyk/issues/310
~Blocked on: https://github.com/runtimeverification/pyk/pull/362~

Most Ethereum verification tools implement Bounded Model Checking, which
is where they unroll each loop a fixed number of times. This improves
their stability to code changes because they don't need to provide loop
invariants.

This PR introduces bounded model checking as a proof technique to the
pyk library:

- Class `AGBMCProof` is introduced, which is an `AGProof`, but it stores
also a list of states that are bounded due to reaching loop-unroll
bound.
- Class `AGBMCProver` is introduced, which is an `AGProver`, but every
iteration of calling the `AGProver` it also checks if any of the new
nodes have reached the depth bound. It accepts the as well a `same_loop`
predicate which allows identifying loop heads for BMC purposes.
- For IMP, the `_is_terminal` check is extended to also consider `<k>
SOME_VAR </k>` as terminal, os that the backend doesn't start inventing
program by narrowing on that case.
- A test of the BMC prover is added which is not true in the unbounded
case, but is true with bounded loop unrolling.
- A test of the BMC prover is added which fails in the code following
the unrolled loop.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>