Support for functional claims via EqualityProof and EqualityProver (#372)
This sets up a specialized prover for equality/functional proofs, which
can be discharged much faster than reachability proofs. The motivation
for this comes from the fact that we are currently using the
reachability prover for KEVM equality proofs (via the "runLemma =>
doneLemma" style proofs), and it's quite a bit slower than it should be
(many calls to RPC server instead of just one call to simplify).

- A simple test of an equality proof for IMP is added.
- The class `EqualityProof` is added, which stores as data all the
components of the equality, as well as the simplified state of the
equality. In addition, a convenience method `EqualityProof.from_claim`
is provided for extracting `EqualityProof` from a `KClaim`.
- The class `EqualityProver` is added which holds an `EqualityProof` and
is able to "advance" it by calling the simplifier on the underlying
equality. If it makes progress, the simplified equality is stored as
part of the proof state.

---------

Co-authored-by: devops <devops@runtimeverification.com>