A conversion KRule -> Axiom (#342)
I tried the output of `_krule_to_kore` to match the output of
```sh
kompile --disable-kore-antileft simple-proofs.k --backend haskell
```
but currently there is a difference: the pattern generated by `kompile`
captures the whole `<state>` cell as a variable which is then reused on
the rhs, while the same rule obtained from `kprint.definition` does not
do that. I haven't explored this issue yet. @ehildenb Is it a problem?

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>