Add KCFG function for computing path constraints (#366)
This PR adds a new function `path_constraints(str) -> KInner` to the
`KCFG` class, which takes a node ID and computes the constraints added
at split edges along the shortest path to that node. If there are any
cover edges along the path, it will apply the cover's substitution to
the collected constraint to make sure the constraints are in terms of
the original variables.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>