Organize KCFG pretty-printing functionality (#403)
This moves all the KCFG pretty-printing functionality over to KCFGShow
class. As a consequence, the `KCFG` datastructure no longer ever needs
access to a `KPrint`!

In particular:

- `KCFG.to_dot` becomes `KCFGShow.dot`
- `KCFG.pretty_segments` and `KCFG.pretty` become
`KCFGShow.pretty_segments` and `KCFGShow.pretty`
- `KCFG.Edge.pretty` and `KCFG.Cover.pretty` are inlined into
`KCFGShow.pretty_segments` as sub-functions `_print_edge` and
`_print_cover`. This mirrors the subfunction `_print_node`.

This does not attempt to make any semantic improvements to the code
there, just move it to the new module and adjust imports and such.

---------

Co-authored-by: devops <devops@runtimeverification.com>