Add `implication_failure_reason` to KCFGExplore (https://github.com/runtimeverification/pyk/pull/415)
Adds the method `implication_failure_reason` to KCFGExplore. This method
was originally written by @ehildenb as part of the K summarizer. It
makes an implication check between two CTerms and attempts to show why
the implication fails, if it fails (the cells that didn't match or the
leftover implication between constraints).
Adds a test
[rc/tests/integration/kcfg/test_impl_failure_reason.py](https://github.com/runtimeverification/pyk/compare/noah/implication-failure-reason?expand=1#diff-f0f5b7f4cadcc3620c868a84af0daf7754f2016b4f66ea2fcbcb0fd2182481a9)
to exercise this method.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>
Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>