Cleanup IMP proof definitions (https://github.com/runtimeverification/pyk/pull/421)
This PR cleans up the IMP proof definitions now that we have an
`_is_terminal` predicate to use for them. Because of this, we don't need
to artificially halt execution with a `#FINISH` rule.

- Remove the `#FINISH` production from `IMP-VERIFICATION`, remove the
`IMP-VERIFICATION.halt` rule altogether, and from the test-suite.
- Rename a variable in the `IMP-SIMPLE-SPEC` that was unused to an
anonymous variable `_:Bool`.
- Fix a bug in `rename_generated_vars` which was dropping the sort
annotation on anonymous variables when renaming them. This is
accomplished by allowing the user to specify a sort when calling
`abstract_term_safely`.
- Disables using the ephemeral runners because they seem to not exist
anymore...

---------

Co-authored-by: devops <devops@runtimeverification.com>