Equality proof updates (https://github.com/runtimeverification/pyk/pull/419)
This PR is factored-out from #394. Changes:

* `EqualityProver.advance_proof` now calls `cterm_implies` to check the
implication of form `(dummy_config #And constraints) #Implies
((dummy_config #And LHS) #Equals (dummy_config #And RHS)))` where
`dummy_config` is the configuration of the semantics (in our case, IMP)
with all cells holding symbolic variables.
* tests for various edge-cases, specifically when the constraints are
`#Bottom`. Is this case, the proof should pass trivially, issuing a
warning to the user.

---------

Co-authored-by: devops <devops@runtimeverification.com>