Ensure no zero-length loops are created in KCFG (https://github.com/runtimeverification/pyk/pull/428)
Part of: https://github.com/runtimeverification/pyk/issues/378

This ensures that no zero-length loops are created in the KCFG by the
user, which will catch all sorts of silly things that could happen when
adding new edges to the KCFG.

- Function `path_length` is factored out of
https://github.com/runtimeverification/pyk/pull/395.
- Function `shortest_path_between` is changed to use `path_length` to
determine the shortest path, rather than just the number of successors.
- A check is added at all the edge-creation sites that the new edge will
not create a zero-length loop.

---------

Co-authored-by: Jan Tusil <jenda.tusil@gmail.com>
Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>