Add pretty printing of nonterminal labels and test (https://github.com/runtimeverification/pyk/pull/438)
Closes #437.
Currently we use the same `pretty_print` function both for our normal
user-facing printing but also for printing claims into k files. This
causes issues because the parser for claims does not recognize the
nonterminal labels. The approach I'm using right now to solve this
problem is to have a flag for the `KPrint` object for whether or not the
labels should be printed, and it's off by default.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>