kprove.py: Pass Haskell RTS options through GHCRTS environment variable (#455)
Previously in #451, `KProve.prove(...)` was modified to pass all haskell
backend arguments through `--haskell-backend-command` in order to
support RTS options.

However, this is problematic because `--debugger` cannot be used with
`--haskell-backend-command`.

To resolve this, we partially revert #451, using `KORE_EXEC_OPTS` for
the common case and ~~only falling back to `--haskell-backend-command`
when RTS options are actually provided~~ using `GHCRTS` for RTS options.

---------

Co-authored-by: devops <devops@runtimeverification.com>