Implement very simple sort parameter inference (https://github.com/runtimeverification/pyk/pull/464)
Related to: https://github.com/runtimeverification/k/issues/3436

This implements very simple sort parameter inference for the Kast ->
Kore conversion in pyk. This inference does not attempt to be clever or
smart in any way, it just handles the simplest case:

- All the arguments in parametric sort positions can have their sort
determined directly.
- All the arguments in a sort parametric position get the same sort as
any other occurrences of that sort parameter.
- All sort parameters get assigned a concrete sort.

This PR:

- Adds two tests of this failing (specifically on `#ite` production),
which were crashing before.
- Adds `KDefinition.sort: KInner -> KSort | None`, which best effort
returns the sort of the supplied term.
- Makes a small refactoring to `KDefinition.sort_vars_subst` to call out
the simple method `greatest_common_subsort`.
- Changes `kast_to_kore` to first call `KDefinition.add_sort_parameters`
before converting to Kore.

---------

Co-authored-by: devops <devops@runtimeverification.com>