Move PrettyPrinter class into kast submodule (https://github.com/runtimeverification/pyk/pull/466)
This is factored out of:
https://github.com/runtimeverification/pyk/pull/433

This PR moves the `PrettyPrinter` class into its own module
`kast.pretty`, as well as addressing some minor printing issues. In
particular:

- Class `PrettyPrinter` is moved from `ktool.kprint` to `kast.pretty`,
and associated helpers are moved too.
- `PrettyPrinter` now takes more arguments associated with unparsing
directly in Python (but not with unparsing using `kast`), including:
  - `definition: KDefinition`: used to build a `SymbolTable`.
- `extra_unparsing_modules: Iterable[KFlatModule]`: used for defining
new symbols on the fly and being able to unparse them.
- `patch_symbol_table: Callable[[SymbolTable], None]`: instead of
relying on overriding the classmethod, it's passed as explicit argument.
- All the appropriate arguments are threaded through `KPrint`, `KRun`,
and `KProve` to be able to build a `PrettyPrinter`.
- `PrettyPrinter.print` is broken into two separate methods:
`PrettyPrinter._print_kouter` and `PrettyPrint._print_kinner`.
  - First it checks which method it should call on the passed in term.
- Teh sub-methods `PrettyPrinter._print_*` are calling the more speciic
methods most of the time, *except*:
- When we switch from printing `KOuter` to printing `KInner` or
vice-versa, we go back through `PrettyPrinter.print` top-level method.
This allows us to do things like call `undo_aliases` _only once_ when we
go to unparse a `KInner`, but also have it work inside a `KRule` when
unparsing the rule body or requires clause, for example.
- The method `pretty_print_kast` is removed, and relevant tests of that
deleted. We no longer expose the functionality of printing directly from
a `SymbolTable`, and that becomes a more opaque internal data-structure.

---------

Co-authored-by: devops <devops@runtimeverification.com>