Allow creation of `EqualityProof` with `proof_dir` using `from_claim` (https://github.com/runtimeverification/pyk/pull/475)
In
[KEVM](https://github.com/runtimeverification/evm-semantics/pull/1860),
EqualityProofs weren't saving because there was no `proof_dir` being
specified. We use `from_claim` to create `EqualityProof`s and currently
this always results in the `proof_dir` being `None`.
This PR adds this as an optional parameter to `from_claim` to be passed
into the `EqualityProof` constructor.

I've tested locally that with this change and the appropriate change to
KEVM we can now store the EqualityProofs to disk and they are not
repeated on subsequent runs.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>