Fix sort params not being inferred correctly during EqualityProof generation (https://github.com/runtimeverification/pyk/pull/481)
We have a function `KDefinition.add_sort_params()` which adds sort
parameters to a KInner term, but this was not being applied in one case
where it was needed, the `sort_strict()` call in `EqualityProof`
generation.
This PR adds a test to demonstrate the issue and fixes it by making sure
the sort params are added before calling `sort_strict()`.

---------

Co-authored-by: devops <devops@runtimeverification.com>