APRProof specific proof state visualization (#489)
Fixes: #413 

This PR introduces proof-specific viewers, and does a few refactors to
make that possible:

- Helpers are added to `KCFG` for turning into modules, including:
- `KCFG.Split.covers` which returns each individual split as a reverse
cover.
- `KCFG.NDBranch.edges`, which returns each non-deterministic step as an
edge.
- `KCFG.Edge.to_rule`, which converts an edge into a K rule (or claim).
- Small refactorings to `KCFGShow` structure:
- `KCFGShow.dot` returns a `Digraph` directly (for further manipulation)
instead of a list of `str`.
- `KCFGShow.to_module` which returns a `KFlatModule` is factored out and
used directly in `KCFGShow.show`.
- Helper methods embedded in other methods in `KCFGShow` are factored
out directly to staticmethods of `KCFGShow` so they can be reused
easily.
- We just treat `stuck` as a node attribute now, printed at the
beginning of a node, instead of printing it a second time following a
node too.
- The way that `KCFGShow` calculates node attributes is adjusted and
factor out.
- A new class `NodePrinter` is factored out, which by default just uses
the attributes the `KCFG` knows about for printing out nodes.
- That class is extended by `APRProofNodePrinter` and
`APRBMCProofNodePrinter`, which add the additional information about the
nodes `pending,terminal,bounded` state. If a node is `terminal` or
`bounded`, then the `stuck` attribute is no longer printed for less
confusion.
- `KCFGShow` now takes an optional `*NodePrinter` as an argument (or
builds on itself), so instantiating that class is how you can customize
how nodes are displayed in text.
- The tests are adapted to use the more specific `*NodePrinter` classes,
and the output checked that the more specific node information is indeed
being included.
- Adds helper predicates `APRProof.is_{pending,terminal}` and
`APRProof.is_bounded`, which should behave better in case there are
qualified node ids passed in.

---------

Co-authored-by: devops <devops@runtimeverification.com>