Implement APRProofViewer as an APRProof specific TUI (https://github.com/runtimeverification/pyk/pull/507)
Following https://github.com/runtimeverification/pyk/pull/489, it turns
out that KEVM's TUI is broken because it relies on the old way of doing
Node printing.

This makes a new `APRProofViewer` class, which largely inherits from
`KCFGViewer`, but adjusts it to handle the more spceific node viewers.

I have tested this with both KEVM+Foundry and raw KEVM specs, and it
seems to work.

---------

Co-authored-by: devops <devops@runtimeverification.com>