Fix LHS and RHS extraction from rules. (https://github.com/runtimeverification/pyk/pull/502)
When normalizing a rewrite, pyk represents something like "(a => .K) ~>
stuff" as "(a => #EmptyK) ~> stuff". When extracting the RHS of such a
rewrite, pyk produces "#EmptyK ~> stuff" which breaks various tools.

This PR replaces "#EmptyK" with an actual empty KSequence (i.e.,
KSequence([])), which, for the example above, will produce just "stuff"
as the RHS.

---------

Co-authored-by: devops <devops@runtimeverification.com>