Support Circularities in `APRProver` (#395)
This pull request enables support for using dependencies during a proof
search done by `APRProver`, as well as support for circularities. The
class `APRProof` now takes a list of dependencies of a claim (other than
the claim itself), plus a flag saying whether the claim is a circularity
- that is, whether it depends on itself. The class `APRProver` then
performs all symbolic execution in a newly-added module (with name
suffix `-DEPENDS-MODULE`) containing the dependencies of the claim
excluding the claim itself; if the claim is a circularity, then all
symbolic execution after one step is performed in another module (with
name suffix `-CIRCULARITIES-MODULE`) containing all the dependencies AND
the claim itself.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>