Adjust BMC depth bound calculation (https://github.com/runtimeverification/pyk/pull/511)
Currently, for `--bmc-depth` set to 0, we won't execute _any_ code. In
addition, for `--bmc-depth 1`, we'll execute a loop a single time, but
then cap it immediately instead of considering the path where we execute
past the end of the loop.

This changes our behavior (and the tests) to make it so that we do _one
more_ iteration of each loop before capping it. The last iteration is
capped before being considered for verification, but the iteration at
the depth bound allows stepping out of the loop to continue the
verification.

This also modifies the `APRProofNodePrinter` to include the
`init/target` nodes in the decorators for printing those KCFGs.

---------

Co-authored-by: devops <devops@runtimeverification.com>