Adjust BMC loop detection for languages with coinciding loop heads and branches (#512)
In GOTO style languages (which use `JUMP` instructions for control
flow), the loop head and the branch points coincide. One example of such
a language is EVM. In these languages, when detecting prior loop heads
for BMC loop detection, we were counting each loop-head twice, because
we were simply walking backwards in the KCFG and looking for nodes that
were in the `_same_loop_head` heuristic relation. But because these loop
heads were also branches, we were counting both the pre- and post-branch
nodes as loop heads, even though it should only count a single time.

To fix this, we basically don't count nodes which are "zero-distance"
from each other as distinct loop heads, which results in us only keeping
the pre-branch node as a loop head.

In particular, we:

- Add some helpers to KCFG/APRProof path detection/manipulation:
- `APRProof.shortest_path_to` is a shortcut for
`KCFG.shortest_path_between(APRProof.init, SOME_NODE)`, utilizing the
fact that every node in a `APRProof` should be guaranteed to be
reachable from the `init` node.
- KCFG.shortest_distance_between` is added as a helper which returns
either `None` (no path between), or the shortest distance between two
nodes (in either diretion).
- `KCFG.zero_distance_between` is a function that returns `True` if
there is a zero-length path between two nodes.
- `APRProof.nonzero_depth` is refactored to use
`KCFG.zero_distance_between`.
- The prior-loop detection for `APRBMCProver` is adjusted so that it:
- Calculates an actual path in order from the root to the current node,
instead of just looking at prior reachable nodes.
- Only includes nodes in that path that are (i) not zero-depth from any
of the other nodes in the current path, and (ii) not zero-depth from the
current node. This prevents double-counting nodes.
- Adds a test of the correct loop-head detection:
- The language `GOTO` is added, inspired by `EVM`, with trimmed down
functionality.
- A proof of the program `sum(N)` is added in language `GOTO`, and a BMC
test harness is added for this proof, showing that it achieves the
correct final branching factor of 2 instead of ending early because it
over-detects loops.

---------

Co-authored-by: devops <devops@runtimeverification.com>