Node refutation and subproofs (https://github.com/runtimeverification/pyk/pull/394)
https://github.com/runtimeverification/pyk/pull/396 Work towards
[evm-semantics#1753](https://github.com/runtimeverification/evm-semantics/issues/1753)

Blocked on: #419 

Changes:

**Subproofs**

* `Proof` now holds a list of proof ids that are required to pass for
the parent proof to pass. If a proof has subproofs, it must have a
`proof_dir` set, and the subproofs are looked up there.
* factor-out `read_proof` to be a standalone function, rather than a
`Proof` static method. This allows looking up the proof type in the JSON
and returning the proof of the appropriate type
* Modifies the `*Proof.status` property to consider subproofs
* adds unit tests for the above logic

Subproofs are intended to serve as "proof obligations", i.e. auxiliary
proofs that are factored-out of a larger ongoing proof.

All subproofs are "direct dependencies" of their parent proof, i.e. the
parent proof cannot have *explicit* chains of subproofs. These chains
could be implicit, i.e. the subproofs can have their own subproofs.

Related work is the `ProofSchedule` (PR #396) --- it is intended to make
transitive dependencies explicit by organizing them into a DAG.

**Node refutation**

Having subproofs allows us to add a possibility to refute nodes in the
`APRProver`. Refuting a node means (see `APRProver.refute_node`):

* marking it as expanded in the KCFG of the proof
* generating a `RefutationProof` that states the following: the
conjunction of the node's conditions along the path AND the the
condition of the first node down the split is unsatisfiable
* Add the generated proof as a subproof of the parent `APRProof`.
* If all stuck nodes of a proof haver a passing refutation, the parent
proof passes.

*Note*: The node refutations proofs are proofs of the unsatisfiability
of the node's path constraint. They have different meaning than
`EqualityProof`'s, and are constructed in a slightly different way. I've
decided to separate `RefutationProof` into its own class that does not
inherit `EqualityProof` to make this distinction clear.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>
Co-authored-by: Noah Watson <noah@nwatson.xyz>
Co-authored-by: Noah Watson <107630091+nwatson22@users.noreply.github.com>