Add `failure_info` method for showing why `APRProof` failed (#517)
Closes #463

This method was previously in `evm-semantics` and this upstreams it into
`pyk`.
Adds a class `APRFailureInfo` which stores the IDs of the pending nodes,
and the IDs of the failing nodes, along with the failure reason
(`implication_failure_reason`) and path condition (`path_constraints`).
Also adds a test which checks the output of this function for
`IMP-SIMPLE-SPEC.failing-if`.

---------

Co-authored-by: devops <devops@runtimeverification.com>