Improve `anti_unify_with_constraints` (#522)
Needed for
https://github.com/runtimeverification/evm-semantics/pull/1934.
Adds an additional option to `anti_unify_with_constraints` which adds an
additional constraint which is an `_orBool_` of the two anti-unifying
substitutions.
Adds a test for the method with this option which tests a new variable
is created to abstract and the `orBool` constraint is added as
described.

---------

Co-authored-by: devops <devops@runtimeverification.com>