Add tests and bug reports for failing`get-model` with `chop` (https://github.com/runtimeverification/pyk/pull/520)
Blocked on:
https://github.com/runtimeverification/haskell-backend/issues/3616

This PR introduces two tests for the `get-model` endpoint that involve
`chop` function defined as:

```
  syntax Int ::= chop ( Int )           [function, total, klabel(chop), smtlib(chop)]
  syntax Int ::= "pow256"               [alias] /* 2 ^Int 256 */

  rule pow256 => 115792089237316195423570985008687907853269984665640564039457584007913129639936
  rule chop ( I:Int ) => I modInt pow256 [concrete, smt-lemma]
```

The first test `sat-chop-single-int` shows that the variable
substitution returned for `chop(x) == x && x == 1` is correct: `x = 1`.
The second, failing, test `sat-chop-sub-ints` shows that the reported
substitution (`x = 0`) is incorrect for a slightly more complex
expression involving subtraction: `chop(1 - x) == 0`.

The changes in
[pyk/testing/_kompiler.py](https://github.com/runtimeverification/pyk/compare/smt-lemma-translation?expand=1#diff-f7eb769a6f82214d26847367701f3229182d9ae9890962becdd1b5af638b3750)
add bug-report generation functionality for tests defined in
`KoreClientTest`.

---------

Co-authored-by: devops <devops@runtimeverification.com>