Add `KoreServerPool` (https://github.com/runtimeverification/pyk/pull/542)
Adds a class `KoreServerPool` that wraps a `ThreadPoolExecutor` and
provides a `KoreServer` instance per thread. As `ThreadPoolExecutor`
handles threads economically (i.e. it won't start a new thread if
there's a free thread that has been started earlier), only the minimal
necessary number of servers will be started (up to `max_workers`).

The pool can be submitted functions that have `int` as first parameter.
The server pool will forward the handler to the thread pool with the
corresponding server port number substituted for that parameter.

---------

Co-authored-by: devops <devops@runtimeverification.com>