Profiling test for loading large semantics, minor optimizations (https://github.com/runtimeverification/pyk/pull/556)
This PR is inspired by the discussion here:
https://github.com/runtimeverification/pyk/issues/548

In particular:

- A profiling test of loading a large JSON Kast definition is added.
This is a definition created from a Foundry test-suite which just has 2
contracts in it. This semantics is ~40MB, when loading the entire actual
KEVM foundry test-suite, it's closer to ~120MB.
- After inspecting the above profiling output, we also drop some input
validation on `KVariable` which is likely not needed, and gets us
something like 10% speedup in converting a definition from JSON -> KAST.
- Some INFO log events are added for crude tracking of where time is
being spent.
- DEBUG log events in `kast_to_kore` and `kore_to_kast` are removed,
which are actually quite expensive even if not turned on (~ 1% of
execution time). These messages haven't been useful for a while, since
our conversions haven't had any new bugs discovered for a while.

---------

Co-authored-by: devops <devops@runtimeverification.com>