Add methods to reachability proof classes for removing nodes (https://github.com/runtimeverification/pyk/pull/561)
Currently the process needed to cleanly remove nodes from a proof is
messy, as it has to be done in multiple places. The method to prune the
node (remove all children except the target node) is on the KCFG, and
when that happens the proof may maintain certain references to those
nodes that no longer exist, e.g. `_terminal_nodes` and `_bounded_nodes`,
which would have to be removed separately.

This PR adds at @ehildenb's suggestion methods that do node removal at
the proof level.

https://github.com/runtimeverification/evm-semantics/pull/1978

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>