Inject into `KItem` instead of `K` (https://github.com/runtimeverification/pyk/pull/569)
closes #565 
We weren't doing the kast to kore conversion in the same way as
`kore-exec`, resulting in a possibly invalid kore term (per discussion
on Slack involving @goodlyrottenapple @tothtamas28 and
@virgil-serbanuta).
This changes the conversion so that the injection is not done directly
into the `K` sort, but into a `KItem` which is embedded into a K
sequence.
I don't know much about kore so I'm not sure if this is correct, but it
did produce the correct expected output on #565.

---------

Co-authored-by: devops <devops@runtimeverification.com>