add function to assert equality between ExecuteResults (https://github.com/runtimeverification/pyk/pull/593)
Better fix for #579 

We need to be able to compare next_states as if it were a set and not a
tuple. This was already being done in one unit test but I wanted a more
general solution. This solution is adapted from the one @tothtamas28
suggested in the linked PR.

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>