Make anti-unification more sort-aware (https://github.com/runtimeverification/pyk/pull/598)
- Changed `anti_unify` and `anti_unify_with_constraints` to require a
`KDefinition` which they use to grab the tightest sort for the terms
they are abstracting
- Since these actions are now associated with a specific K definition,
moved them into `KDefinition` as methods.
- Adds `KDefinition.least_common_supersort` which returns the most
specific sort that will cover two terms (as long as one sort is a
subsort of the other).
- Uses this function in `KDefinition.sort` to determine the sort of a
`KRewrite`.
- Rewrote the tests for `anti_unify` and `anti_unify_with_constraints`
to use `KDefinition` and test that the abstracted variables are now more
tightly typed based on the original terms.

This is an attempt to fix an issue in
https://github.com/runtimeverification/evm-semantics/pull/1934 where the
looser sort bound is preventing simplification after the merging
(anti-unification) of two nodes.

---------

Co-authored-by: devops <devops@runtimeverification.com>