print no-junk with flat multi-or (#3588)
The Maude backend Kore-RPC server doesn't start with as much stack space
as kompile, and this was leading it to not have enough stack space to
parse the definition.kore that was generated by kompile, primarily
because of the deeply nested \or patterns. In general, we want to try to
avoid situations with potentially arbitrarily deeply nested KORE because
it makes things more difficult for recursive descent parsers.

Luckily, we already have a solution to this problem. The llvm and
haskell backend both support the \left-assoc symbol to be used in
conjunction with \or, which allows a flat multi-or-like representation
that still ultimately parses to the same term in the end. This allows
the recursive descent parser to parse the axioms in question with a
constant amount of stack space, no matter how many productions or sorts
exist in the definition.