KCFGExploration: KCFG + terminal nodes (https://github.com/runtimeverification/pyk/pull/601)
This PR introduces the `KCFGExploration` class, which encapsulates the
`KCFG` class with information about `terminal` and `explorable` nodes
and associated reasoning. The main changes to existing classes, in
addition to the adaptation of using `KCFGExploration` instead of `KCFG`,
are as follows:

- In `KCFGExplore`, the `extend` function now checks automatically
whether or not newly added nodes are `terminal`. This removes the need
for these checks inside the provers.
- In `APRProof` and `APRBMCProof`, the notions of `pending` and
`failing` nodes have been slightly reformulated
- In `APRProver` and `APRBMCProver`, subsumption is cached and checked
for`terminal` nodes at the end of every prover iteration

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>
Co-authored-by: Andrei Văcaru <16517508+anvacaru@users.noreply.github.com>