Rule label cannot contain ` or whitespace (#3596)
Adds a check to the compiler that reports an error if a sentence label
contains the backtick character or whitespace. This ought to allow the
Maude backend to compile rule labels directly to Maude identifiers
without some clumsy encoding system.

I am reasonably certain that it should not break anything to add this
restriction.