Testing run nix-flake on self hosted linux machine (#3609)
This PR is a temporary fix to the `error: writing to file: No space left
on device` issue on Nix Flake ubuntu-20.04 CI.

The problem seems to be in nix downloading multiple versions of the same
package. @goodlyrottenapple will investigate it further in the near
future.

This fix is meant only to unlock the CI for the various PRs that need to
be tested with Nix Flake running on a Ubuntu machine.

From @goodlyrottenapple as well: "Doesn't need to be specific ubuntu
version"