Add Bencher to track Benchmarks Performance Regression (#3611)
Fixes #3517
Follow up of #3603

This PR Introduces Bencher.dev to track performance regressions on our
benchmarks.
The Benchmarks already run in the Test-PR Workflow as "Performance
Tests" CI. This PR introduces a third-party program that will collect
the JSON results of these executions and display them in a graphical
online tool.

It will also set a threshold for the time/space that each test, and if
an execution passes this limit, the tool will post a comment on the PR
warning the developer which tests the regression was observed. The CI
will also fail, which won't prevent the PR from being merged.